Adding YAML support for SDK configuration
Fix #41
Fix #40

This PR adds support for YAML format in defining SDK configuration. This is a
required change for integration with EdgeX Foundry as of their new 3.0 release.

I do not like the duplicated boilerplate introduced by this PR and thought for
some time about how I might reduce it. Since I was unable for the moment to
devise an elegant solution, I am leaving this as-is for now in the hopes some
clarity will be forthcoming later.

Two issues I wrestled with in trying to devise a better plan are:
* I do not want to pollute the package with a bunch of globally defined "alias"
  types used during unmarshaling
* The manner in which JSON/YAML are unmarshaled is just different enough to be
  problematic. It would be nice if the interfaces behaved the same. Instead,
  some "flag" parameter would need to be passed around to indicate to the
  unmarshal "wrapper" helper method which should be used.

All unit tests pass.

Signed-off-by: Trevor Conn <trevor_conn@dell.com>